% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF.Echidna.R, R/basF.R
\name{rho.par}
\alias{rho.par}
\alias{AF.base}
\alias{filterD1}
\alias{aov.batch}
\title{Summary of some R basic functions}
\usage{
rho.par(rho,x0,a=1,b=0)

filterD1(x, ..., except = NULL)

aov.batch(df,mod,nf=NULL,alpha=.05)
}
\description{
Summary of some R basic functions
}
\details{
This package would supply some functions for R. Details as following:

\tabular{ll}{
\strong{Function} \tab \strong{Description} \cr
\code{\link{datatable2}}    \tab Create an HTML table widget. \cr
\code{\link{mc.run}}    \tab multi-core run in R. \cr
\code{\link{spd.plot}}    \tab Plot spatial data or Variogram.\cr
\code{\link{Var}}    \tab Output Variance components for R packages.
}
}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
