% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bdR.pin.R
\name{bdR.pin}
\alias{bdR.pin}
\title{Count error for h2 in BreedR}
\usage{
bdR.pin(object, formula, signif = FALSE, digit = 5)
}
\arguments{
\item{object}{an object of BreedR result}

\item{formula}{formula for h2 (or corr, not now)}

\item{signif}{Index to output signif levels, F(default) for non-signif.}

\item{digit}{a number to control decimal point,5(default).}
}
\value{
the result is returned directly.
}
\description{
\code{bdR.pin} This function counts standard error(se) for heritability(h2) and corr value
and also outputs significent level for corr value in BreedR package.
}
\details{
counts standard error(se) for heritability(h2)
}
\examples{
library(breedR)
res.animal <- remlf90(fixed = phe_X ~ 1,
                      random = ~ gg,
                      genetic = list(model = 'add_animal',
                      pedigree = globulus[, 1:3],
                      id = 'self'),
                      data = globulus)
bdR.pin(res.animal, h2~V2/(V1+V2+V3))
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
\references{
bdRPlus website:https://github.com/yzhlinscau/AAfun
}

