% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basF.R
\name{datatable2}
\alias{datatable2}
\title{Create an HTML table widget for dataset}
\usage{
datatable2(data)
}
\arguments{
\item{data}{a data object (either a matrix or a data frame).}
}
\description{
\code{datatable2} This function  creates an HTML widget to display
rectangular data (a matrix or data frame) using the JavaScript
library DataTables.
}
\examples{

library(AFEchidna)
data("iris")
datatable2(iris)

}
