% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF.Echidna.R
\name{Echidna.options}
\alias{Echidna.options}
\title{Model Qualifiers for Echidna}
\description{
Model Qualifiers for Echidna
}
\details{
The Qualifiers control various aspects of the model fitting process and reporting of results:

\cr
\tabular{ll}{
\strong{jobqualf } \tab \strong{Description} \cr
\code{!WORKSPACE n}       \tab where n is an integer between 1 and 32 gigabytes. \cr
\code{!CONTINUE [f] or !FINAL [f] }       \tab instructs Echidna to retrieve variance parameters from an .esv file. \cr
\code{!VIEW }    \tab  displays any Winteracter graphics directly to the screen, such as '!view !PNG' \cr
\code{!DEBUG}          \tab requests additional debugging output to be written. \cr
\code{!LOGFILE}          \tab receive the DEBUG information. \cr
\code{!EQN [q]}          \tab sets the !EQN qualifier which selects the equation ordering option, q could be 1~7 or 77. \cr
\code{others}           \tab !RENAME,!ARGS,!OUTFOLDER, no needs in R. 
}

\cr
\tabular{ll}{
\strong{qualifier} \tab \strong{Description} \cr
\code{!EXTRA n}     \tab forces n more iterations after convergence . \cr
\code{!SINGLE}      \tab forces Echidna not to use Parallel Processing. \cr
\code{!SLOW}      \tab reduces stepsize when updating variance parameters so convergence is slower but
possible more reliable. \cr
\code{!SKIP i}      \tab indicating the file contains i heading lines to be ignored. \cr
\code{!READ f}      \tab specifying that f data fields are to be read. \cr
\code{!FILTER Variable !SELECT value !EXCLUDE value}         \tab subset the data sets. \cr
\code{!MVINCLUDE}   \tab design variables which are missing are assumed to be zero.\cr
\code{!MVREMOVE}    \tab data records are ignored if any design variables have missing values.
}

\cr
\tabular{ll}{
\strong{GRM qualifiers } \tab \strong{Description} \cr
\code{!SKIP i}       \tab to skip the first i lines of the grm/giv file. \cr
\code{!LDET d}       \tab lets you set the logDet value for the supplied giv matrix. \cr
\code{!ADD value}    \tab adds value to the diagonal of a GRM matrix before inversion.value can only be 1, 2 or 3. \cr
\code{!PSD}          \tab allows a GRM matrix to be positive semidefinite. \cr
\code{!NSD}          \tab allows a GRM matrix to be negative semidefinite. \cr
\code{!ND}           \tab allows a GRM matrix to be negative definite. 
}

\cr Data fields: if VARIABLE is a CLASS variable, the NAME must be followed by a qualifier indicating it
is a class variable.

\tabular{ll}{
\strong{class qualifier} \tab \strong{Description} \cr
\code{*}               \tab variable is coded 1:n. \cr
\code{!I n}            \tab variable is coded with INTEGER labels, not 1:n. \cr
\code{!A n}            \tab variable is coded with ALPHANUMERIC labels. \cr
\code{!L <labels>}     \tab data is coded 1:n and <labels> is the list of n class names. \cr
\code{!AS <factor> }   \tab when two or more alphanumeric variables have a common list of class names. \cr
\code{!LL c }          \tab reset the maximum length to c characters. \cr
\code{!PRUNE }         \tab reset the number of levels (n) in a factor to the real maximum levels. 
}

\cr PEDIGREE FILE: ifile contains 3 or 4 fields being the identity (tag, id, name) of
       an individual, its Sire and its Dam..

\tabular{ll}{
\strong{qualifier} \tab \strong{Description} \cr
\code{!SKIP i}        \tab the file contains i heading lines to be ignored. \cr
\code{!CSV}           \tab file is strictly COMMA delimited, without .csv suffix. \cr
\code{!GROUPS g }     \tab first g lines are genetic groups. . \cr
\code{!MGS}           \tab  the third field is a maternal grandsire (not DAM). \cr
\code{!SAVE f }       \tab requests the inverse be written as a .giv (f=1, ascii) or .bgiv (f=3, real binary) file. 
}


\cr VPREDICT statements have the syntax as: <Key Letter> <Label> <arguments>.
   <Key Letter> is one of the letters F, H, R, V, etc.

\tabular{ll}{
\strong{Key Letter } \tab \strong{Description} \cr
\code{F}       \tab specifies a linear function like animal + units. \cr
\code{H}       \tab specifies a ratio (heritability) like animal / Total. \cr
\code{R}       \tab specifies correlations from a matrix or formula. \cr
\code{V}       \tab converts a FA structure to a US structure. \cr 
\code{X}       \tab is a multiply function. \cr
\code{S}       \tab is a square root function. \cr
\code{K}       \tab sets short vectors for constants (e.g. Legendre coefficients) to be used for M. \cr
\code{M}       \tab uses the K coefficients to convert US (us(leg(dim,3))) matrix for a specific dim. \cr
\code{C}       \tab moves components back e.g. C label I[:II]=J[:JJ] where I < J. \cr
\code{D}       \tab discards components eg D from 393. \cr
\code{W}       \tab writes components to f.vpc and their variance to f.vpv.
}

\cr Variance structures

\tabular{ll}{
\strong{function} \tab \strong{Description} \cr
\code{id(),idv()}      \tab identity or scaled identity. \cr
\code{ar1(), ar1v()}   \tab autoregressive correlation or covariance matrix. \cr
\code{coru(), coruv(), coruh()}   \tab uniform correlation matrix, v for common variance,h for heterogeneous variance. \cr
\code{diag()}          \tab diagonal variance matrix. \cr
\code{grmk()}          \tab the kth GRM matrix. \cr
\code{mrmk()}          \tab multiple relationship matrices. \cr
\code{us()}            \tab unstructured variance matrix.\cr 
\code{facvk()}         \tab factor analytic (basic form). \cr
\code{xfak()}          \tab factor analytic (eXtended form). \cr
\code{rrk()}           \tab factor analytic: No specific variance. 
}
}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
