% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fm.R
\docType{data}
\name{fm}
\alias{fm}
\title{fm dataset}
\format{one dataset with 13 variables
\describe{
 \item{TreeID} {Tree Sample code}
 \item{Spacing} {Tree Spacing code}
 \item{Rep} {Replication code}
 \item{Fam} {Family code}
 \item{Plot} {Plot code}
 \item{dj} {junior wood density}
 \item{dm} {mature wood density}
 \item{wd} {total wood density}
 \item{h1} {tree heighth for 1th year}
 \item{h2} {tree heighth for 2th year}
 \item{h3} {tree heighth for 3th year}
 \item{h4} {tree heighth for 4th year}
 \item{h5} {tree heighth for 5th year}
}}
\usage{
fm
}
\description{
Example dataset
}
\keyword{datasets}

