% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fs.R
\docType{data}
\name{fs}
\alias{fs}
\title{full-sib dataset}
\format{one dataset with 8 variables
\describe{
 \item{TreeID} {Tree Sample code}
 \item{Mum} {Parent Mum code}
 \item{Dad} {Parent Dad code}
 \item{Fam} {Family code}
 \item{Fam2} {Family code2}
 \item{Rep} {Replication code}
 \item{Plot} {Plot code}
 \item{dbh10} {Tree DBH at 10th year}
}}
\usage{
fs
}
\description{
full-sib dataset
}
\keyword{datasets}

