% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF.Echidna.R
\name{GenomicRel}
\alias{GenomicRel}
\title{Generate genomic relationship matrix.}
\usage{
GenomicRel(marker,option,ped=NULL,
                   Infv=1000,Gm=NULL,G.adj=FALSE,Gres=TRUE)
}
\arguments{
\item{marker}{markers data.}

\item{option}{option (1~5) for different G matrixs.}

\item{ped}{ped data.}

\item{Infv}{A value for Inf in G matrix.}

\item{Gm}{G matrix from marker.}

\item{G.adj}{Adjust G matrix with A matrix from ped data.}

\item{Gres}{return G matrix directly(True, default) or in data frame(FALSE).}
}
\description{
\code{GenomicRel} This function generates 5 genomic relationship matrixs.

\tabular{ll}{
\strong{option} \tab \strong{Description} \cr
1    \tab observed allele frequencies (GOF, VanRaden, 2008). \cr
2    \tab weighted markers by recipricals of expected variance (GD, Forni et al., 2011). \cr
3    \tab allele frequencies fixed at 0.5 (G05, Forni et al., 2011). \cr
4    \tab allele frequencies fixed at mean for each locus (GMF, Forni et al., 2011).\cr
5    \tab regression of MM' on A sort (Greg, VanRaden, 2008).
}
}
\examples{
\dontrun{
library(AFEchidna)

read.example(package = "AFEchidna", setpath = TRUE)
Markers<-read.file(file="sim_markers.txt",sep=' ' )
ped<-read.table( "sim_pedigree.txt", sep=' ')


 GOF1=GenomicRel( Markers,1)
 GD1=GenomicRel(  Markers,2)
 G051=GenomicRel( Markers,3)
 GMF1=GenomicRel( Markers,4)
 
 # the same result but adjust G matrix with ped data:
 GOF2=GenomicRel(  Markers,1, ped,G.adj=T)
 GD2=GenomicRel(   Markers,2, ped,G.adj=T)
 G052=GenomicRel(  Markers,3, ped,G.adj=T)
 GMF2=GenomicRel(  Markers,4, ped,G.adj=T)
 Greg=GenomicRel(  Markers,5, ped,G.adj=T)
}

}
\references{
Isik Fikret. Genetic data analysis for plant and animal breeding. 2017
}
\author{
Isik Fikret
}
