% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basF.R
\name{mc.run}
\alias{mc.run}
\title{multi-core run in R}
\description{
multi-core run in R
}
\details{
This page would show how to run with multi-core in R.
}
\examples{
\dontrun{

code<-c('600519.ss','601398.ss')

# stock.get is the main function
# using lapply and stock.get
mm<-lapply(code, stock.get)

# using  stock.get with multi-core
library(doParallel)
cl <- makePSOCKcluster(detectCores()-1)
registerDoParallel(cl)

## multi-core run here
mm1<-foreach(code=code) \%dopar\% stock.get(code)

stopCluster(cl)
}
}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
