% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF.Echidna.R
\name{met.plot}
\alias{met.plot}
\alias{met.corr}
\alias{met.biplot}
\alias{met.vmat}
\title{MET analysis}
\usage{
met.plot(data, plot.title = NULL,...)

met.corr(object,siteV,kN=NULL,horiz=TRUE,rotate=FALSE)

met.biplot(object,siteV,biplot=FALSE, dSco.u=NULL,dLam.u=NULL)

met.vmat(object,siteV,VmatN,corN)
}
\arguments{
\item{data}{MET data.}

\item{plot.title}{MET plot title.}

\item{object}{echidna factor analytic results for MET, such as mm.}

\item{siteV}{trial site values, a vector or data frame.}

\item{horiz}{output cluster site result format, horiz(default).}

\item{rotate}{Rotate the factor loadings, FALSE(default).}

\item{biplot}{output biplots, FALSE(default).}

\item{dSco.u}{Least score of Variety breeding value.}

\item{dLam.u}{Least distance from center.}

\item{kn}{Site cluster group numbers, 3(default).}
}
\description{
\code{met.plot} plots MET data.\code{met.corr} calculates var/cov/corr from echidna MET factor analytic 
results to further research the relation of trial sites.
\code{met.biplot} This function biplots MET factor analytic results from echidna 
to find the relation of trial sites and the best variety suitable to trial sites.
}
\examples{
\dontrun{
library(AFEchidna)

path="C:/Users/echi/exam" #home
setwd(path)

MET<-read.csv('MET.csv')
names(MET)

# example 1
# variable order: yield,genotype,site,row,col
MET2<-MET[,c(9,1,2,4:5)] 
str(MET2)
met.plot(MET2)

# example 2
MET3<-MET[,c(9,1,2,4:7)] # add variable order on MET2: Rep, Block
str(MET3)
met.plot(MET3,"My met trials")

## running met analysis with FA model
mm<-echidna(es0.file="MET.es0",trait='yield',fixed='Loc',
      random='Genotype.xfa2(Loc)',
      residual='sat(Loc).units', #sat(Loc).ar1(Col).ar1(Row)
      #predict=c('Genotype'),
      vpredict=c('V Vmat Genotype.xfa1(Loc)','R cor 20:40'),
      qualifier='!maxit 50 !SLN',
      foldN='mm',
      met=T)
      
Var(mm)

siteV<-unique(MET['Loc']) # should be a data.frame or vector

met.corr(mm,siteV=siteV)

met.biplot(mm,siteV=siteV)
met.biplot(mm,siteV=siteV,biplot=T)
met.biplot(mm,siteV=siteV,biplot=T,dSco=1.0,dLam=0.8)

res2<-met.vmat(mm,siteV=siteV,VmatN='Vmat',corN='cor')
res2$res
res2$var

}

}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
