% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF.Echidna.R
\name{model.comp}
\alias{model.comp}
\title{Model comparison for Echidna.}
\usage{
model.comp(...,LRT=FALSE,boundary = TRUE)
}
\arguments{
\item{...}{A list with more than two Echidna-R results, such as "m1,m2,m3,m4".}

\item{LRT}{Value TRUE for Likelihood ratio test (LRT), default (FALSE) for no LRT.}

\item{boundary}{boundary If \code{TRUE} (the default) hypothesized parameter
values being tested lie on the boundary of the parameter space.}
}
\description{
\code{model.comp} This function would compare models with different random structure 
under the same fixed factors.
}
\examples{

\dontrun{
library(AFEchidna)



              
#####   model comparison    #####
model.comp(m1,m2)
model.comp(m1,m2,LRT=TRUE)
}

}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
