% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basF.R
\name{PC.res}
\alias{PC.res}
\title{output PCs for AMMI analysis}
\usage{
PC.res(model, pcN = 2)
}
\arguments{
\item{model}{an AMMI object.}

\item{pcN}{keeping pc number (default, 2).}
}
\description{
\code{PC.res} This function outputs PCs for AMMI analysis.
}
\examples{

library(AFEchidna)
data(plrv,package='agricolae')
names(plrv) #str(plrv)

library(agricolae)
model<- with(plrv,AMMI(ENV=Locality,GEN=Genotype,REP=Rep,
                       Y=Yield,console=F,PC=T))

model$ANOVA # for aov
model$analysis # for pc

PCs.aov<-PC.res(model,2) # for GEI aov
PCs.aov

}
