\name{PrSpa}
\alias{PrSpa}
\docType{data}
\title{
Tree spacing trial dataset
}
\description{
A dataset of spacing trials of Pinus radiata
}
\usage{data(PrSpa)}
\format{
  A data frame with 827 observations on the following 13 variables.
  \describe{
    \item{\code{TreeID}}{a factor with levels \code{80001} \code{80002} \code{80004} \code{80005} \code{80008} \code{80026} \code{80028} \code{80033} \code{80034} \code{80035} \code{80036} \code{80037} \code{80038} \code{80045} \code{80046} \code{80048} \code{80049} \code{80052} \code{80055} \code{80056} \code{80057} \code{80058} \code{80059} \code{80061} \code{80062} \code{80063} \code{80070} \code{80073} \code{80074} \code{80078} \code{80080} \code{80084} \code{80085} \code{80087} \code{80088} \code{80093} \code{80094} \code{80101} \code{80103} \code{80105} \code{80106} \code{80108} \code{80109} \code{80110} \code{80112} \code{80118} \code{80120} \code{80126} \code{80127} \code{80129} \code{80130} \code{80131} \code{80132} \code{80133} \code{80134} \code{80136} \code{80137} \code{80138} \code{80139} \code{80140} \code{80143} \code{80145} \code{80147} \code{80148} \code{80151} \code{80152} \code{80153} \code{80154} \code{80155} \code{80169} \code{80173} \code{80175} \code{80178} \code{80179} \code{80186} \code{80187} \code{80188} \code{80189} \code{80191} \code{80192} \code{80193} \code{80194} \code{80199} \code{80200} \code{80215} \code{80217} \code{80218} \code{80224} \code{80226} \code{80227} \code{80234} \code{80236} \code{80238} \code{80239} \code{80240} \code{80241} \code{80244} \code{80246} \code{80248} \code{80249} \code{80250} \code{80252} \code{80253} \code{80255} \code{80256} \code{80258} \code{80259} \code{80264} \code{80265} \code{80268} \code{80270} \code{80271} \code{80272} \code{80275} \code{80280} \code{80283} \code{80284} \code{80285} \code{80286} \code{80287} \code{80293} \code{80294} \code{80295} \code{80297} \code{80298} \code{80300} \code{80302} \code{80304} \code{80306} \code{80310} \code{80311} \code{80314} \code{80316} \code{80318} \code{80320} \code{80321} \code{80323} \code{80324} \code{80326} \code{80327} \code{80329} \code{80330} \code{80332} \code{80333} \code{80335} \code{80337} \code{80338} \code{80341} \code{80342} \code{80343} \code{80346} \code{80350} \code{80351} \code{80355} \code{80356} \code{80363} \code{80364} \code{80365} \code{80366} \code{80367} \code{80369} \code{80372} \code{80375} \code{80378} \code{80379} \code{80381} \code{80384} \code{80385} \code{80387} \code{80390} \code{80392} \code{80393} \code{80395} \code{80396} \code{80397} \code{80401} \code{80403} \code{80408} \code{80410} \code{80412} \code{80413} \code{80414} \code{80417} \code{80418} \code{80420} \code{80421} \code{80424} \code{80425} \code{80426} \code{80427} \code{80428} \code{80429} \code{80430} \code{80437} \code{80438} \code{80439} \code{80441} \code{80443} \code{80444} \code{80446} \code{80447} \code{80448} \code{80465} \code{80469} \code{80472} \code{80473} \code{80474} \code{80477} \code{80480} \code{80481} \code{80482} \code{80484} \code{80485} \code{80486} \code{80487} \code{80499} \code{80500} \code{80508} \code{80509} \code{80512} \code{80513} \code{80514} \code{80515} \code{80518} \code{80519} \code{80523} \code{80525} \code{80528} \code{80529} \code{80530} \code{80531} \code{80532} \code{80533} \code{80534} \code{80535} \code{80537} \code{80540} \code{80541} \code{80542} \code{80545} \code{80546} \code{80547} \code{80548} \code{80549} \code{80550} \code{80551} \code{80552} \code{80554} \code{80555} \code{80558} \code{80559} \code{80561} \code{80563} \code{80564} \code{80565} \code{80566} \code{80573} \code{80574} \code{80575} \code{80576} \code{80577} \code{80578} \code{80581} \code{80585} \code{80586} \code{80587} \code{80588} \code{80599} \code{80601} \code{80606} \code{80607} \code{80608} \code{80610} \code{80611} \code{80613} \code{80615} \code{80623} \code{80624} \code{80625} \code{80626} \code{80627} \code{80629} \code{80633} \code{80634} \code{80635} \code{80638} \code{80639} \code{80640} \code{80641} \code{80642} \code{80643} \code{80645} \code{80651} \code{80652} \code{80653} \code{80654} \code{80656} \code{80665} \code{80666} \code{80667} \code{80670} \code{80675} \code{80676} \code{80677} \code{80678} \code{80680} \code{80681} \code{80687} \code{80689} \code{80690} \code{80693} \code{80696} \code{80698} \code{80700} \code{80702} \code{80705} \code{80706} \code{80708} \code{80709} \code{80712} \code{80713} \code{80714} \code{80715} \code{80716} \code{80717} \code{80719} \code{80721} \code{80724} \code{80725} \code{80727} \code{80730} \code{80731} \code{80732} \code{80735} \code{80736} \code{80737} \code{80738} \code{80742} \code{80743} \code{80744} \code{80747} \code{80748} \code{80750} \code{80751} \code{80752} \code{80753} \code{80755} \code{80756} \code{80758} \code{80760} \code{80762} \code{80764} \code{80765} \code{80767} \code{80769} \code{80771} \code{80772} \code{80777} \code{80781} \code{80782} \code{80783} \code{80784} \code{80795} \code{80797} \code{80798} \code{80800} \code{80801} \code{80802} \code{80803} \code{80804} \code{80806} \code{80808} \code{80809} \code{80813} \code{80814} \code{80817} \code{80820} \code{80823} \code{80825} \code{80830} \code{80831} \code{80832} \code{80836} \code{80837} \code{80838} \code{80840} \code{80842} \code{80847} \code{80849} \code{80852} \code{80853} \code{80854} \code{80861} \code{80863} \code{80864} \code{80867} \code{80869} \code{80870} \code{80871} \code{80872} \code{80873} \code{80874} \code{80876} \code{80879} \code{80881} \code{80883} \code{80885} \code{80888} \code{80889} \code{80890} \code{80894} \code{80895} \code{80897} \code{80898} \code{80901} \code{80904} \code{80908} \code{80909} \code{80911} \code{80913} \code{80915} \code{80916} \code{80921} \code{80922} \code{80924} \code{80929} \code{80930} \code{80932} \code{80933} \code{80935} \code{80937} \code{80945} \code{80947} \code{80948} \code{80951} \code{80952} \code{80953} \code{80954} \code{80955} \code{80957} \code{80961} \code{80962} \code{80963} \code{80965} \code{80967} \code{80968} \code{80969} \code{80971} \code{80973} \code{80974} \code{80976} \code{80977} \code{80980} \code{80981} \code{80983} \code{80986} \code{80991} \code{80992} \code{80997} \code{80998} \code{81001} \code{81002} \code{81003} \code{81004} \code{81005} \code{81009} \code{81010} \code{81014} \code{81015} \code{81016} \code{81017} \code{81021} \code{81022} \code{81024} \code{81029} \code{81030} \code{81031} \code{81033} \code{81034} \code{81035} \code{81039} \code{81041} \code{81042} \code{81043} \code{81045} \code{81046} \code{81047} \code{81050} \code{81051} \code{81052} \code{81053} \code{81054} \code{81055} \code{81057} \code{81060} \code{81061} \code{81062} \code{81063} \code{81064} \code{81069} \code{81070} \code{81071} \code{81075} \code{81081} \code{81082} \code{81083} \code{81085} \code{81087} \code{81089} \code{81094} \code{81099} \code{81100} \code{81105} \code{81110} \code{81112} \code{81113} \code{81114} \code{81116} \code{81118} \code{81123} \code{81124} \code{81125} \code{81128} \code{81130} \code{81131} \code{81133} \code{81134} \code{81136} \code{81137} \code{81138} \code{81139} \code{81141} \code{81146} \code{81148} \code{81150} \code{81152} \code{81153} \code{81155} \code{81156} \code{81162} \code{81165} \code{81167} \code{81168} \code{81169} \code{81171} \code{81172} \code{81173} \code{81174} \code{81179} \code{81180} \code{81181} \code{81183} \code{81184} \code{81186} \code{81187} \code{81191} \code{81193} \code{81197} \code{81200} \code{83604} \code{83608} \code{83609} \code{83613} \code{83624} \code{83628} \code{83632} \code{83633} \code{83636} \code{83637} \code{83638} \code{83639} \code{83642} \code{83651} \code{83652} \code{83655} \code{83656} \code{83658} \code{83660} \code{83661} \code{83663} \code{83668} \code{83671} \code{83681} \code{83682} \code{83696} \code{83698} \code{83706} \code{83708} \code{83709} \code{83710} \code{83711} \code{83712} \code{83713} \code{83714} \code{83721} \code{83722} \code{83723} \code{83729} \code{83732} \code{83736} \code{83737} \code{83762} \code{83773} \code{83774} \code{83777} \code{83778} \code{83779} \code{83786} \code{83788} \code{83797} \code{83798} \code{83800} \code{83804} \code{83808} \code{83811} \code{83815} \code{83816} \code{83818} \code{83825} \code{83827} \code{83832} \code{83833} \code{83838} \code{83840} \code{83842} \code{83850} \code{83853} \code{83857} \code{83859} \code{83860} \code{83861} \code{83863} \code{83870} \code{83873} \code{83874} \code{83875} \code{83898} \code{83899} \code{83900} \code{83902} \code{83910} \code{83917} \code{83934} \code{83936} \code{83937} \code{83945} \code{83948} \code{83953} \code{83957} \code{83960} \code{83961} \code{83963} \code{83964} \code{83968} \code{83975} \code{83979} \code{83980} \code{83982} \code{83984} \code{83989} \code{83992} \code{83993} \code{83996} \code{83998} \code{83999} \code{84001} \code{84002} \code{84003} \code{84006} \code{84010} \code{84012} \code{84018} \code{84022} \code{84024} \code{84027} \code{84028} \code{84029} \code{84038} \code{84043} \code{84045} \code{84047} \code{84057} \code{84065} \code{84091} \code{84093} \code{84106} \code{84114} \code{84123} \code{84126} \code{84128} \code{84130} \code{84140} \code{84142} \code{84147} \code{84151} \code{84154} \code{84159} \code{84161} \code{84170} \code{84172} \code{84174} \code{84178} \code{84180} \code{84185} \code{84187} \code{84192} \code{84194} \code{84196} \code{84198} \code{84202} \code{84210} \code{84214} \code{84215} \code{84216} \code{84221} \code{84224} \code{84225} \code{84232} \code{84233} \code{84238} \code{84240} \code{84241} \code{84244} \code{84245} \code{84249} \code{84254} \code{84260} \code{84263} \code{84267} \code{84287} \code{84311} \code{84342} \code{84345} \code{84352} \code{84359} \code{84364} \code{84365} \code{84366} \code{84368} \code{84372} \code{84377} \code{84379} \code{84389} \code{84391} \code{84392} \code{84398} \code{84409} \code{84410} \code{84412} \code{84414} \code{84415} \code{84416} \code{84421} \code{84422} \code{84424} \code{84428} \code{84430} \code{84441} \code{84454} \code{84455} \code{84480} \code{84485} \code{84510} \code{84512} \code{84522} \code{84523} \code{84529} \code{84530} \code{84546} \code{84547} \code{84548} \code{84550} \code{84555} \code{84561} \code{84564} \code{84565} \code{84568} \code{84569} \code{84572} \code{84582} \code{84583} \code{84585} \code{84587} \code{84593} \code{84600} \code{84603} \code{84605} \code{84606} \code{84607} \code{84608} \code{84609} \code{84619} \code{84621} \code{84625} \code{84644} \code{84645} \code{84647} \code{84658} \code{84669} \code{84672} \code{84674} \code{84675} \code{84686} \code{84707} \code{84710} \code{84711} \code{84715} \code{84717} \code{84728} \code{84731} \code{84738} \code{84739} \code{84747} \code{84755} \code{84758} \code{84759} \code{84761} \code{84762} \code{84776} \code{84780} \code{84781} \code{84786} \code{84788} \code{84794} \code{84795} \code{84797} \code{84799}}
    \item{\code{Spacing}}{a factor with levels \code{2} \code{3}}
    \item{\code{Rep}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5}}
    \item{\code{Fam}}{a factor with levels \code{70001} \code{70002} \code{70003} \code{70004} \code{70005} \code{70006} \code{70007} \code{70008} \code{70009} \code{70010} \code{70011} \code{70012} \code{70015} \code{70016} \code{70017} \code{70018} \code{70019} \code{70020} \code{70021} \code{70022} \code{70023} \code{70024} \code{70025} \code{70027} \code{70028} \code{70029} \code{70030} \code{70031} \code{70032} \code{70033} \code{70034} \code{70035} \code{70036} \code{70037} \code{70038} \code{70039} \code{70040} \code{70041} \code{70043} \code{70044} \code{70045} \code{70046} \code{70047} \code{70048} \code{70050} \code{70051} \code{70052} \code{70053} \code{70054} \code{70055} \code{70056} \code{70057} \code{70059} \code{70060} \code{70061}}
    \item{\code{Plot}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
    \item{\code{dj}}{a numeric vector}
    \item{\code{dm}}{a numeric vector}
    \item{\code{wd}}{a numeric vector}
    \item{\code{h1}}{a numeric vector}
    \item{\code{h2}}{a numeric vector}
    \item{\code{h3}}{a numeric vector}
    \item{\code{h4}}{a numeric vector}
    \item{\code{h5}}{a numeric vector}
  }
}

\references{
Yuanzhen Lin, Xiaoyang Chen. R & ASReml-R Statistical Analysis Tutorial. China Forestry Publishing House. 2014
}
\examples{
data(PrSpa)
## maybe str(PrSpa) ; plot(PrSpa) ...
}
\keyword{datasets}
