% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF.Echidna.R
\name{summary}
\alias{summary}
\title{Summarize an esR object}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{An \code{esR} object.}
}
\value{
A list of class \code{summary.esR} with the following components:

\describe{

\item{org.res}{Original results from .esr file in Echidna.} 
  
\item{varcomp}{A dataframe summarising the random variance component.}

\item{IC}{nedf, loglik, Akaike information criterion and Bayesian information criterion.}

\item{coef.fixed}{A dataframe of coefficients and their standard errors
for fixed effects.}

\item{coef.random}{A dataframe of coefficients and their standard errors
for random effects.}

}
}
\description{
A \code{summary} method for objects inheriting from class
\code{esR}.
}
