% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF.Echidna.R, R/basF.R
\name{Var.esR}
\alias{Var.esR}
\alias{Var}
\alias{Var.lme}
\alias{Var.lmerMod}
\alias{Var.asreml}
\alias{Var.remlf90}
\title{Output Variance components for R packages}
\usage{
\method{Var}{esR}(object)

Var(object, ...)

\method{Var}{lme}(object)

\method{Var}{lmerMod}(object)

\method{Var}{asreml}(object)

\method{Var}{remlf90}(object, mulT = FALSE)
}
\arguments{
\item{object}{an object of mixed model results from R packages.}

\item{mulT}{multi-trait model(default, FALSE).}
}
\value{
the result is returned directly.
}
\description{
\code{Var} This function output Variance components for R packages.
}
\details{
Output Variance component for mixed model results from R packages.
}
\examples{
library(AFEchidna)

## 00 data
data(butron.maize,package='agridat')
df<-butron.maize

## 01 nlme package
library(nlme) # V3.1-131

nlm<-lme(yield~1+env,random=~1|male/female,
               na.action='na.omit',
               data=df)

Var(nlm)

## 02 lme4 package
library(lme4) # V1.1-17

lme<-lmer(yield~1+env+(1|male)+(1|female),
               data=df)

Var(lme)

 \dontrun{
## 03 breedR package
library(breedR) # V0.12-1

bdR <- remlf90(fixed = yield~1+env,
                      random = ~ male+female,
                      data=df)

Var(bdR)


## 04 asreml package
library(asreml) #V3.0

asr <- asreml(fixed = yield~1+env,
                      random = ~ male+female,
                      na.method.X='include',
                      data=df)

Var(asr)

##### special for breedR
library(breedR)

data(globulus)
res.animal <- remlf90(fixed = phe_X ~ 1,
                      random = ~ gg,
                      genetic = list(model = 'add_animal',
                      pedigree = globulus[, 1:3],
                      id = 'self'),
                      data = globulus)
Var(res.animal)
}


}
\references{
AFfR website:https://github.com/yzhlinscau/AFfR
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
