% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fs.R
\docType{data}
\name{zinc}
\alias{zinc}
\title{zinc dataset}
\format{one dataset with 3 variables
\describe{
  \item{Family}{Family code}
  \item{SeedZn}{Seed Zinc content}
  \item{LeafZn}{Leaf Zinc content}
}}
\usage{
zinc
}
\description{
zinc dataset
}
\keyword{datasets}

