% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF.Echidna.R
\name{pin}
\alias{pin}
\alias{pin.esR}
\alias{pin11}
\alias{pin22}
\title{Count error for h2 and corr.}
\usage{
pin(
  object,
  mulp = NULL,
  idx = 1,
  digit = 3,
  all = FALSE,
  signif = FALSE,
  Rres = FALSE
)

\method{pin}{esR}(
  object,
  mulp = NULL,
  idx = 1,
  digit = 3,
  all = FALSE,
  signif = FALSE,
  Rres = FALSE
)

pin11(object,formula=NULL,idx=1,signif=FALSE,
                digit=3,all=FALSE,Rres=FALSE)

pin22(...,mulp,signif=FALSE,digit=3)
}
\arguments{
\item{object}{Echidna result object in R.}

\item{mulp}{multi-formula for h2, corr, etc.}

\item{idx}{trait order to run, 1,2...}

\item{digit}{Index for decimal number, 3(default).}

\item{all}{Show variance component and genetic parmameter together, FALSE(default).}

\item{signif}{Index to output signif levels, FALSE(default) for non-signif.}

\item{Rres}{Index(TRUE) to restore results, FALSE(default).}

\item{formula}{formula for h2 or corr.}
}
\value{
the result is returned directly.
}
\description{
\code{pin} This function counts standard error(se) for heritability(h2) and
corr value and also outputs significent level for corr value in asreml and breedR package.
}
\details{
Count error for h2 and corr value, also outputs significent level.
}
\examples{
\dontrun{

 library(AFEchidna)

 AFEchidna::read.example(package='AFEchidna', setpath=TRUE)
 file.edit("fm.es0")
 
 res<-echidna(h3~1+Rep,
              random=~Fam,es0.file="fm.es0")
 
 # pin results if using vpredict function
 pin(res)
 
 # run pin function to count genetic parameters
 pin11(res,h2~V1/(V1+V2))
 pin(res,mulp=c(h2~V1/(V1+V2),h2f~V1/(V1+V2/4))
}
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
